remove most of the SVG code
looks like we won't be using SVG
remove XML from libraries to link
add sanitized filename for output
